streaming-visualization/requirements.txt
python-dotenv
geojson
geopandas
streamlit
folium
streamlit-folium
contextily
shapely
confluent-kafka